requirements.txt
2captcha-python==1.1.3
beepy==1.0.7
websocket-client==1.5.0
